Refactor import paths for firestore in Matches and Sports pages
